feat: sistema de upload de imagens concluído

+++ catalogo/views.py
@@ -1,14 +1,13 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
 from django.http import Http404
-
 from .models import Filme
 from .forms import FilmeForm
 
 @login_required
 def lista_filmes(request):
     """Exibe todos os filmes cadastrados."""
-    filmes_qs = Filme.objects.order_by('-data_cadastro')
+    filmes_qs = Filme.objects.filter(usuario=request.user).order_by('-data_cadastro')
     contexto = {'filmes': filmes_qs}
     return render(request, 'catalogo/lista_filmes.html', contexto)
 
@@ -18,9 +17,11 @@ def novo_filme(request):
     if request.method != 'POST':
         form = FilmeForm()
     else:
-        form = FilmeForm(data=request.POST)
+        form = FilmeForm(request.POST, request.FILES)
         if form.is_valid():
-            form.save()
+            novo_filme_obj = form.save(commit=False)
+            novo_filme_obj.usuario = request.user
+            novo_filme_obj.save()
             return redirect('catalogo:lista_filmes')
     
     contexto = {'form': form}
@@ -30,20 +31,27 @@ def novo_filme(request):
 def editar_filme(request, filme_id):
     """Edita um filme existente."""
     filme = Filme.objects.get(id=filme_id)
+    if filme.usuario != request.user:
+        raise Http404
+
     if request.method != 'POST':
         form = FilmeForm(instance=filme)
     else:
-        form = FilmeForm(instance=filme, data=request.POST)
+        form = FilmeForm(request.POST, request.FILES, instance=filme)
         if form.is_valid():
             form.save()
             return redirect('catalogo:lista_filmes')
-    context = {'filme': filme, 'form': form}
-    return render(request, 'catalogo/editar_filme.html', context)
+
+    contexto = {'filme': filme, 'form': form}
+    return render(request, 'catalogo/editar_filme.html', contexto)
 
 @login_required
 def apagar_filme(request, filme_id):
     """Apaga um filme existente."""
     filme = Filme.objects.get(id=filme_id)
+    if filme.usuario != request.user:
+        raise Http404
+    
     if request.method == 'POST':
         filme.delete()
         return redirect('catalogo:lista_filmes')
